move create_sample_run into test_utils
